Crate pallet_plasm_rewards

Plasm rewards Module

The Plasm rewards module provides functionality for handling whole rewards and era.

plasm_rewards::Config

Call

Module

Overview

The Plasm staking module puts together the management and compensation payment logic of the ERA. The Plasm Rewards module calls the Dapps Staking and Validator. It also allocates rewards to each module according to the Plasm Token Ecosystem inflation model.

Re-exports

pub use traits::*;

pub use $crate::traits::GetPalletVersion as _;

Modules

inflation
This is explains plasm inflation models. The staking has 2 kinds.

traits

Structs

ActiveEra
The active era information, it holds index and start.

ActiveEraInfo
Information regarding the active era (era in used in session).

BondedEras
A mapping from still-bonded eras to the first session index of that era.

CurrentEra
The current era index.

ErasStartSessionIndex
The session index at which the era start for the last HISTORY_DEPTH eras

ForDappsEraReward
This is the compensation paid for the dapps operator of the Plasm Network. This is stored on a per-era basis.

ForSecurityEraReward
This is the compensation paid for the security of the Plasm Network. This is stored on a per-era basis.

ForceEra
True if the next session change will be a new era regardless of index.

GenesisConfig
Genesis config for the module, allow to build genesis storage.

HistoryDepth
Number of era to keep in history.

Module

Enums

Call
Dispatchable calls.

Error
Error for the staking module.

Forcing
Mode of era-forcing.

RawEvent
Events for this module.

Releases

Traits

Config

Type Definitions

BalanceOf

EraIndex

Event
RawEvent specialized for the configuration Config

MomentOf

SessionIndex
Simple index type with which we can count sessions.