Trait pallet_plasm_rewards::Trait

pub trait Trait: Trait {
    type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>;
    type Time: Time;
    type SessionsPerEra: Get<SessionIndex>;
    type BondingDuration: Get<EraIndex>;
    type ComputeEraForDapps: ComputeEraWithParam<EraIndex>;
    type ComputeEraForSecurity: ComputeEraWithParam<EraIndex>;
    type ComputeTotalPayout: ComputeTotalPayout<<Self::ComputeEraForSecurity as ComputeEraWithParam<EraIndex>>::Param, <Self::ComputeEraForDapps as ComputeEraWithParam<EraIndex>>::Param>;
    type MaybeValidators: MaybeValidators<EraIndex, Self::AccountId>;
    type Event: From<Event<Self>> + Into<Self::Event>;
}

Associated Types

type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>

The staking balance.

type Time: Time

Time used for computing era duration.

type SessionsPerEra: Get<SessionIndex>

Number of sessions per era.

type BondingDuration: Get<EraIndex>

Number of eras that staked funds must remain bonded for.

type ComputeEraForDapps: ComputeEraWithParam<EraIndex>

Get the amount of staking for dapps per era.

type ComputeEraForSecurity: ComputeEraWithParam<EraIndex>

Get the amount of staking for security per era.

type ComputeTotalPayout: ComputeTotalPayout<<Self::ComputeEraForSecurity as ComputeEraWithParam<EraIndex>>::Param, <Self::ComputeEraForDapps as ComputeEraWithParam<EraIndex>>::Param>

How to compute total issue PLM for rewards.

type MaybeValidators: MaybeValidators<EraIndex, Self::AccountId>

Maybe next validators.

type Event: From<Event<Self>> + Into<Self::Event>

The overarching event type.

Loading content...