Trait pallet_plasm_rewards::traits::ComputeTotalPayout

pub trait ComputeTotalPayout<ValidatorParam, DappsParam> {
    fn compute<N, M>(
        total_tokens: N, 
        era_duration: M, 
        for_security_parm: ValidatorParam, 
        for_dapps_param: DappsParam
    ) -> (N, N)
    where
        N: BaseArithmetic + Clone + From<u32>,
        M: BaseArithmetic + Clone + From<u32>;
}

The reward is allocated from the total supply of tokens, the time for Era, the amount of staking for Security, and the amount of staking for Dapps.

Required methods

fn compute<N, M>(
total_tokens: N,
era_duration: M,
for_security_parm: ValidatorParam,
for_dapps_param: DappsParam
) -> (N, N) where
N: BaseArithmetic + Clone + From<u32>,
M: BaseArithmetic + Clone + From<u32>,

Loading content...