Trait pallet_plasm_rewards::traits::EraFinder

pub trait EraFinder<EraIndex, SessionIndex, Moment> {
    fn current() -> Option<EraIndex>;

    fn active() -> Option<ActiveEraInfo<Moment>>;

    fn start_session_index(era: &EraIndex) -> Option<SessionIndex>;
}

Get the era for validator and dapps staking module.

Required methods

fn current() -> Option<EraIndex>

The current era index.

This is the latest planned era, depending on how session module queues the validator set, it might be active or not.

fn active() -> Option<ActiveEraInfo<Moment>>

The active era information, it holds index and start.

The active era is the era currently rewarded. Validator set of this era must be equal to SessionInterface::validators.

fn start_session_index(era: &EraIndex) -> Option<SessionIndex>

The session index at which the era start for the last HISTORY_DEPTH eras

Loading content...